Put auth stash cache/origin on port 1094, unauth on 1095
ATLAS and CMS use auth and the default XRootD port of 1094.  With the
goal of only shipping a single XCache package that serves all three
data federations, we should unify our ports.

Additionally, we're unifying the HTTP/S and XRootD ports in
osg-xrootd, hence the change from 8443.

This will require changes to clients such as stashcp.