Update 40-stash-cache-plugin.cfg
Re: pss.ccmlib, this was a suggestion from Wei on June 26, 2019:

I add an extra line "pss.ccmlib /usr/lib64/XrdName2NameDCP4RUCIO.so" to the config file.  It might be needed in the future. I expect the current Xcache will ignore it. If it causes trouble, please comment it out.

We should have this in only if/when needed and we know what it does.