Set `GridMapfile` and `GmapOpt` in the XRootD config  (SOFTWARE-5468)
These are variables that an upcoming version of osg-xrootd will use to
define where the grid-mapfile is (instead of hardcoding
`/etc/grid-security/grid-mapfile`) and what the behavior should be if
there's no mapping (instead of hardcoding `trymap`). It has no effect
with previous versions of osg-xrootd.

It has to come before 50-osg-http.cfg and 50-osg-xrdvoms.cfg.